requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor==0.8.1
bottle==0.12.18
certifi==2020.6.20
chatterbot-corpus==1.2.0
click==7.1.2
cmake==3.18.0
coverage==5.2.1
cycler==0.10.0
Django==3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
joblib==0.16.0
mysqlclient==2.0.1
numpy==1.19.1
opencv-contrib-python==4.4.0.40
opencv-python==4.4.0.40
pandas==1.1.0
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
wincertstore==0.2
